pull out common functionality
