+++ scripts/versioning/check-version.ts
@@ -1,34 +1,30 @@
 // this script will be triggered from workflow actions when a new tag is push
-
-import {ethers, run} from 'hardhat'
+import {ethers} from 'hardhat'
 import {Contract} from 'ethers'
 import {ERC20Treasury} from '../../typechain'
 import {log} from '../../config/logging'
+import {deployContract} from '../common'
 
 const AddressZero = ethers.constants.AddressZero
 
 // for each deployable contract, check that the hardcoded version matches the tag version
 async function checkReleaseTag(
-    gitReleaseTag: string,
     contractName: string
 ): Promise<boolean> {
-    await run('compile') // think run compiles by default - remove?
-
-    const sourceTag = process.env.SOURCE_TAG
-    if(sourceTag){
-        log.info(`git tag from workflow: ${sourceTag}`)
+    const gitSourceTag = process.env.SOURCE_TAG
+    if(gitSourceTag){
+        log.info(`git tag from workflow: ${gitSourceTag}`)
     }
     else {
         log.info(`git tag from workflow not defined!`)
-        // panic?
+        process.exit(1)
     }
     
-
     log.info(
-        `Checking ${contractName} release against git release tag ${gitReleaseTag}`
+        `Checking ${contractName} release against git release tag ${gitSourceTag}`
     )
 
-    // how do we generically deploy contracts?
+    // how do we generically deploy contracts, with various constructor params
     const contract: Contract = await deployContract<ERC20Treasury>(
         contractName,
         AddressZero,
@@ -37,40 +33,21 @@ async function checkReleaseTag(
     log.info(`deployed contact at ${contract.address}`)
 
     // eslint-disable-next-line @typescript-eslint/no-unsafe-call
-    const contractVersion: unknown = await contract.getVersion()
-    log.info(`Contract version: ${contractVersion as string}`)
-    if (gitReleaseTag === contractVersion) {
+    const contractVersion: string = (await contract.getVersion() as unknown) as string
+    log.info(`Contract version: ${contractVersion}`)
+    if (gitSourceTag === contractVersion) {
         return true
     }
-    return false
+    throw new Error(`Contract version mismatch! ${contractVersion} vs. ${gitSourceTag}`)
 }
 
 log.info('running checkReleaseTag()')
-checkReleaseTag('v0.0.1', 'ERC20Treasury')
+checkReleaseTag('ERC20Treasury')
     .then((resolve) => {
         log.info(resolve)
-        log.info('success')
         process.exit(0)
     })
     .catch((error) => {
         log.error(error)
         process.exit(1)
     })
-
-// common with scripts/bond/deploy.ts --refactor?
-async function deployContract<T extends DeployableContract<T>>(
-    name: string,
-    ...args: Array<unknown>
-): Promise<T> {
-    const factory = await ethers.getContractFactory(name)
-    const contract = <T>(<unknown>await factory.deploy(...args))
-
-    log.info('%s deployed to: %s', name, contract.address)
-
-    return contract.deployed()
-}
-
-interface DeployableContract<T> {
-    deployed(): Promise<T>
-    address: string
-}
